+++ contracts/Acknowledger.sol
@@ -4,6 +4,7 @@ contract Acknowledger {
     event AcknowledgeFoo(uint256 a);
     event AcknowledgeBarSingle(uint256 a);
     event AcknowledgeBarDouble(uint256 a, uint256 b);
+    event AcknowledgeFallback();
 
     function foo(uint256 a) public {
         emit AcknowledgeFoo(a);
@@ -15,5 +16,9 @@ contract Acknowledger {
 
     function bar(uint256 a, uint256 b) public {
         emit AcknowledgeBarDouble(a, b);
+    }
+
+    function () external {
+        emit AcknowledgeFallback();
     }
 }