+++ src/send.js
@@ -2,9 +2,11 @@ const ethjsABI = require('ethjs-abi');
 
 function findMethod (abi, name, args) {
   for (let i = 0; i < abi.length; i++) {
-    const methodArgs = abi[i].inputs.map(input => input.type).join(',');
-    if ((abi[i].name === name) && (methodArgs === args)) {
-      return abi[i];
+    if (abi[i].type === 'function') {
+      const methodArgs = abi[i].inputs.map(input => input.type).join(',');
+      if ((abi[i].name === name) && (methodArgs === args)) {
+        return abi[i];
+      }
     }
   }
 }
@@ -14,6 +16,7 @@ async function transaction (target, name, argsTypes, argsValues, opts = {}) {
   const encodedData = ethjsABI.encodeMethod(abiMethod, argsValues);
 
   opts.from = opts.from || (await web3.eth.getAccounts())[0];
+
   return web3.eth.sendTransaction({ data: encodedData, to: target.address, ...opts });
 }